async await works to populate the arrays
